+++ SideMenu.podspec
@@ -29,7 +29,7 @@ Pod::Spec.new do |s|
   # s.social_media_url = 'https://twitter.com/<TWITTER_USERNAME>'
 
   s.ios.deployment_target = '10.0'
-  s.swift_version = '4.2'
+  s.swift_version = '5'
 
   s.source_files = 'Pod/Classes/**/*'
   # s.resource_bundles = {
