--- SideMenuPresentable.swift
@@ -1,18 +1,0 @@
-//
-//  SideMenuPresentable.swift
-//  SideMenu
-//
-//  Created by Jon Kent on 12/29/18.
-//
-
-import UIKit
-
-//typealias SideMenuPresentableController = UIViewController & SideMenuPresentable
-//
-//public protocol SideMenuPresentable {
-//    var leftSide: Bool { get }
-//    var menuPresentStyle: MenuPresentStyle { get }
-//    var menuWidth: CGFloat { get }
-//    var menuAlwaysAnimate: Bool { get }
-//    var menuDismissOnPush: Bool { get }
-//}
